Sync Dev with Main (#18)
* [DB] Add Transactions (#12)

* cs# version bump

* add transaction method

* save cps with transaction;dot for millis

co-pilot = op

* small patch

* Change hook methods for OnTriggerStart/EndTouch

* don't start timer after `!s <X>`

* reset `IsStageMode`

* Latest "update" and reworks (#17)

* hi

Summary of completed items:
- Resolved hibernation issues with timers.
- Implemented async queries and methods with multiple database connections.
- Moved PlayerProfile and Map queries to respective classes.
- Added ReplayFrame Situations for situation-specific replay handling.
- Corrected Map Replays reload and ReplayFrameSituations usage.
- Updated Checkpoints comparison messages for correct display.
- Adjusted Stage Replays tracking, saving, and reloading during map runs.
- Fixed checkpoint saving on first map run.
- Refer to TODO for more information

* bonus replays

* readme + todo

* Update README.md

* Update README.md

* Update README.md

Fix `TODO` link

---------

+++ src/ST-Player/PlayerStats/Checkpoint.cs
@@ -3,10 +3,10 @@ namespace SurfTimer;
 internal class Checkpoint : PersonalBest
 {
     public int CP { get; set; }
-    public float EndTouch { get; set; }
+    public int EndTouch { get; set; }
     public int Attempts { get; set; }
 
-    public Checkpoint(int cp, int ticks, float startVelX, float startVelY, float startVelZ, float endVelX, float endVelY, float endVelZ, float endTouch, int attempts)
+    public Checkpoint(int cp, int ticks, float startVelX, float startVelY, float startVelZ, float endVelX, float endVelY, float endVelZ, int endTouch, int attempts)
     {
         CP = cp;
         Ticks = ticks; // To-do: this was supposed to be the ticks but that is used for run_time for HUD????
